+++ src/test/java/it/maggioli/web/rest/EmailPersonaResourceIT.java
@@ -40,14 +40,14 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @WithMockUser
 public class EmailPersonaResourceIT {
 
-    private static final Integer DEFAULT_ID_PERSONA_REF = 1;
-    private static final Integer UPDATED_ID_PERSONA_REF = 2;
+    private static final Long DEFAULT_ID_PERSONA_REF = 1L;
+    private static final Long UPDATED_ID_PERSONA_REF = 2L;
 
-    private static final Integer DEFAULT_ETICHETTA = 1;
-    private static final Integer UPDATED_ETICHETTA = 2;
+    private static final Long DEFAULT_ETICHETTA = 1L;
+    private static final Long UPDATED_ETICHETTA = 2L;
 
-    private static final Integer DEFAULT_NUMERO = 1;
-    private static final Integer UPDATED_NUMERO = 2;
+    private static final Long DEFAULT_NUMERO = 1L;
+    private static final Long UPDATED_NUMERO = 2L;
 
     @Autowired
     private EmailPersonaRepository emailPersonaRepository;
@@ -184,9 +184,9 @@ public class EmailPersonaResourceIT {
             .andExpect(status().isOk())
             .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
             .andExpect(jsonPath("$.[*].id").value(hasItem(emailPersona.getId().intValue())))
-            .andExpect(jsonPath("$.[*].idPersonaRef").value(hasItem(DEFAULT_ID_PERSONA_REF)))
-            .andExpect(jsonPath("$.[*].etichetta").value(hasItem(DEFAULT_ETICHETTA)))
-            .andExpect(jsonPath("$.[*].numero").value(hasItem(DEFAULT_NUMERO)));
+            .andExpect(jsonPath("$.[*].idPersonaRef").value(hasItem(DEFAULT_ID_PERSONA_REF.intValue())))
+            .andExpect(jsonPath("$.[*].etichetta").value(hasItem(DEFAULT_ETICHETTA.intValue())))
+            .andExpect(jsonPath("$.[*].numero").value(hasItem(DEFAULT_NUMERO.intValue())));
     }
     
     @Test
@@ -200,9 +200,9 @@ public class EmailPersonaResourceIT {
             .andExpect(status().isOk())
             .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
             .andExpect(jsonPath("$.id").value(emailPersona.getId().intValue()))
-            .andExpect(jsonPath("$.idPersonaRef").value(DEFAULT_ID_PERSONA_REF))
-            .andExpect(jsonPath("$.etichetta").value(DEFAULT_ETICHETTA))
-            .andExpect(jsonPath("$.numero").value(DEFAULT_NUMERO));
+            .andExpect(jsonPath("$.idPersonaRef").value(DEFAULT_ID_PERSONA_REF.intValue()))
+            .andExpect(jsonPath("$.etichetta").value(DEFAULT_ETICHETTA.intValue()))
+            .andExpect(jsonPath("$.numero").value(DEFAULT_NUMERO.intValue()));
     }
     @Test
     @Transactional
@@ -304,8 +304,8 @@ public class EmailPersonaResourceIT {
             .andExpect(status().isOk())
             .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
             .andExpect(jsonPath("$.[*].id").value(hasItem(emailPersona.getId().intValue())))
-            .andExpect(jsonPath("$.[*].idPersonaRef").value(hasItem(DEFAULT_ID_PERSONA_REF)))
-            .andExpect(jsonPath("$.[*].etichetta").value(hasItem(DEFAULT_ETICHETTA)))
-            .andExpect(jsonPath("$.[*].numero").value(hasItem(DEFAULT_NUMERO)));
+            .andExpect(jsonPath("$.[*].idPersonaRef").value(hasItem(DEFAULT_ID_PERSONA_REF.intValue())))
+            .andExpect(jsonPath("$.[*].etichetta").value(hasItem(DEFAULT_ETICHETTA.intValue())))
+            .andExpect(jsonPath("$.[*].numero").value(hasItem(DEFAULT_NUMERO.intValue())));
     }
 }
